refactor(cubesql): Remove unexpected re-exporting in the crate::
